Create helper validatePage to do the validation and store some common utilities functions
